+++ sdk/go/equinix/doc.go
@@ -1,0 +1,3 @@
+// A Pulumi package for creating and managing equinix-metal cloud resources.
+//
+package equinix