+++ CHANGELOG.md
@@ -2,5 +2,6 @@ CHANGELOG
 =========
 
 ## HEAD (Unreleased)
+* Initial creation of the provider against v1.0.0 of the Equinix Metal Terraform Provider.
 
 ---